Move the code to copy the native DLL into it's own class (#5).
This can be shared with the sample projects - not ideal, but better than not running.
